+++ eslint.config.js
@@ -1,10 +1,8 @@
 import { defineConfig } from 'eslint/config';
-import mochaPlugin from 'eslint-plugin-mocha';
 import js from '@eslint/js';
 import globals from 'globals';
 
 export default defineConfig([
-    mochaPlugin.configs.recommended,
     {
         files: ['**/*.js'],
         plugins: {
@@ -21,7 +19,6 @@ export default defineConfig([
         ],
         rules: {
             'comma-dangle': 'off',
-            'mocha/no-mocha-arrows': 'off',
             indent: ['error', 4],
             quotes: ['error', 'single'],
         },